Handling when is rerun, in anycase, stow can be used, but I wanted to do my own version

+++ script_dotfiles.sh
@@ -60,11 +60,17 @@ function setSymlink(){
   local lastFileName
   lastFileName="$(basename "$dirFile")"
   if [[ "$typeFile" =~ $DOTFILE_HOME_CONFIG  ]]; then
-    ln -s "$dirFile" "$SYMLINK_HOME/$lastFileName"
+    if [[ ! -L "$SYMLINK_HOME/$lastFileName" ]]; then
+      ln -s "$dirFile" "$SYMLINK_HOME/$lastFileName"
+    fi
   elif [[ "$typeFile" =~ $DOTFILE_CONFIG_CONFIG  ]]; then
+    if [[ ! -L "$SYMLINK_CONFIG/$lastFileName" ]]; then
     ln -s "$dirFile" "$SYMLINK_CONFIG/$lastFileName"
+    fi
   else 
+    if [[ ! -L  "$SYMLINK_PICTURES/$lastFileName" ]]; then
      ln -s "$dirFile" "$SYMLINK_PICTURES/$lastFileName"
+    fi
   fi
 }
 function setDotfiles(){
@@ -89,6 +95,7 @@ function setDotfiles(){
       for ((j=0;j<filesInDirectoryLength;j+=1));do
         local resultType
         resultType="$(getType "${dotfilesArray[i]}")"
+
         setSymlink "${filesInDirectory[j]}" "$resultType"
       done
     fi
